main: have only one file in working directory at all times
Fixes some (but not all) blanking some may be experiencing in kde
plasma.  To convince main.qml that the otherFile is different, use a
different file name instead of endlessly cycling between two images.
Because currentImage is already loaded, we only need to supply a valid
otherFile in the time period of the kde slideshow, which should be
really quick inside the /dev/shm directory.

There is likely an even more correct way to do this. (use a pipe or some
such thing instead, and only delete the old image after the new one is
read?)