main: implement parameter parsing and process reuse
Raise window of previous process when starting a second.  Do this by
running a local server that accepts filenames passed on its socket by
lines. (the first line is working directory context, so don't forget it!)
Removes need for a lockfile and the ugly message at the beginning when
launching the program while it is already running.

The practical upshot means that a .desktop placed on the panel will
change the wallpaper when a filed is dropped on to it.